feat(mvvm): Refactor Restaurant Model
