--- solutions/devsprint-vinicius-carvalho-1/DeliveryAppChallenge/Models/RestaurantsListModel.swift
@@ -1,24 +1,0 @@
-//
-//  RestaurantsListModel.swift
-//  DeliveryAppChallenge
-//
-//  Created by Alley Pereira on 24/01/22.
-//
-
-import Foundation
-
-struct RestaurantsListModel: Codable {
-	let name: String
-	let category: String
-	let deliveryTime: DeliveryTime
-
-	struct DeliveryTime: Codable {
-		let min: Int
-		let max: Int
-	}
-
-	private enum CodingKeys: String, CodingKey {
-		case name, category
-		case deliveryTime = "delivery_time"
-	}
-}
